Rename package to dev.saurabharora
Renamed because of the need to publish to mavenCentral which needs us to verify our domain ownership and I own the saurabharora.dev domain
