+++ app/build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 31
     defaultConfig {
-        applicationId "com.saurabharora.lint"
+        applicationId "dev.saurabharora.lint"
         minSdkVersion 21
         targetSdkVersion 31
         versionCode 1
